Fix bad references after rebase due to package renaming
