+++ base/src/test/java/uk/ac/ebi/atlas/solr/EmbeddedSolrCollectionProxyFactory.java
@@ -5,7 +5,8 @@ import org.apache.solr.client.solrj.embedded.EmbeddedSolrServer;
 import org.apache.solr.core.CoreContainer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import uk.ac.ebi.atlas.solr.cloud.fullanalytics.AnalyticsCollectionProxy;
+import uk.ac.ebi.atlas.solr.cloud.collections.AnalyticsCollectionProxy;
+
 
 import java.io.IOException;
 import java.nio.file.Files;
